Fix(Win11Theming): Adding Opt-in for New styles and controls
